Split lib.rs: extract accept.rs and util.rs
Move content negotiation (ScenegraphFormat, preferred_scenegraph_format,
Accept-header parsing) into a new accept module, and
error/header/font/log/version helpers (error_response,
append_vlc_logs_header, format_log_entries, parse_google_font_args,
vegalite_versions) plus CommonOpts/validate_common_opts into a new util
module. Public API preserved via pub re-exports; handler modules'
super:: imports unchanged. 2 accept tests moved into accept.rs alongside
the code they cover.
